requirements.txt
future==0.18.2
glfw==1.8.3
moderngl==5.5.3
multipledispatch==0.6.0
numpy==1.17.3
Pillow==6.2.2
PyOpenGL==3.1.0
six==1.13.0
PyGLM==1.1.2
simpleaudio=1.0.4
